+++ cisco_sdwan/tasks/implementation/_report.py
@@ -43,6 +43,7 @@ class SectionModel(BaseModel):
 
 
 class ReportContentModel(BaseModel):
+    metadata: Optional[dict] = None
     globals: Optional[dict[str, Any]] = None
     sections: list[SectionModel]
 
@@ -118,12 +119,15 @@ def load_content_spec(spec_file: Optional[str], spec_json: Optional[str],
 class Report:
     DEFAULT_SUBSECTION_NAME = "Default"
 
-    def __init__(self, filename: str, section_dict: Optional[dict] = None) -> None:
+    def __init__(self, filename: str, section_dict: Optional[dict] = None, metadata: Optional[dict] = None) -> None:
         # Report section_dict is {<section_name>: {<subsection_name>: [<subsection lines]}}
         self.section_dict = {} if section_dict is None else section_dict
         self.filename = filename
+        self.section_json = []
+        self.metadata = metadata
 
     def add_section(self, section_name: str, subsection_list: list) -> None:
+        self.section_json.extend(json.loads(subsection.json()) for subsection in subsection_list)
         for subsection in subsection_list:
             if isinstance(subsection, Table):
                 subsection_name = subsection.name or Report.DEFAULT_SUBSECTION_NAME
@@ -193,6 +197,11 @@ class Report:
         }
         return Report(self.filename, trimmed_section_dict)
 
+    def save_json(self, filename: str) -> None:
+        report_json = {"metadata": self.metadata, "data": self.section_json}
+        with open(filename, "w") as target_f:
+            target_f.write(json.dumps(report_json, indent=4))
+
 
 @TaskOptions.register('report')
 class TaskReport(Task):
@@ -214,7 +223,8 @@ class TaskReport(Task):
                                    help='report from the specified directory instead of target vManage')
         create_parser.add_argument('--diff', metavar='<filename>', type=existing_file_type,
                                    help='generate diff between the specified previous report and the current report')
-
+        create_parser.add_argument('--save-json', metavar='<filename>', type=filename_type,
+                                   help='export report as JSON-formatted file')
         diff_parser = sub_tasks.add_parser('diff', help='generate diff between two reports')
         diff_parser.set_defaults(subtask_handler=TaskReport.subtask_diff)
         diff_parser.add_argument('report_a', metavar='<report a>', type=existing_file_type,
@@ -249,7 +259,7 @@ class TaskReport(Task):
         self.log_info("Loading report specification")
         content_spec = load_content_spec(parsed_args.spec_file, parsed_args.spec_json, DEFAULT_CONTENT_SPEC)
 
-        report = Report(parsed_args.file)
+        report = Report(parsed_args.file, metadata=content_spec.metadata)
         for description, task_cls, task_args in self.section_iter(content_spec, api is not None, parsed_args.workdir):
             try:
                 task_output = task_cls().runner(task_args, api)
@@ -270,6 +280,9 @@ class TaskReport(Task):
         # Saving current report after running the diff in case the previous report had the same filename
         report.save()
         self.log_info(f'Report saved as "{parsed_args.file}"')
+        if parsed_args.save_json:
+            report.save_json(parsed_args.save_json)
+            self.log_info(f'JSON-formatted report saved as "{parsed_args.save_json}"')
 
         return result
 
@@ -323,7 +336,7 @@ class TaskReport(Task):
             spec_section_args = section.args or {}
             if section.inherit_globals:
                 imported_global_args = {
-                    k: v for k, v in spec_global_args.items() if k in task_meta.task_args_cls.__fields__
+                    k: v for k, v in spec_global_args.items() if k in task_meta.task_args_cls.model_fields
                 }
                 spec_args = {**imported_global_args, **spec_section_args}
             else:
@@ -413,6 +426,7 @@ class ReportCreateArgs(TaskArgs):
     file: Annotated[str, Field(validate_default=True)] = None
     workdir: Optional[str] = None
     diff: Optional[str] = None
+    save_json: Optional[str] = None
     spec_file: Optional[str] = None
     spec_json: Optional[str] = None
 
@@ -420,6 +434,7 @@ class ReportCreateArgs(TaskArgs):
     _validate_workdir = field_validator('workdir')(validate_workdir)
     _validate_existing_file = field_validator('spec_file', 'diff')(validate_existing_file)
     _validate_json = field_validator('spec_json')(validate_json)
+    _validate_filename = field_validator('save_json')(validate_filename)
 
     @field_validator('file', mode='before')
     @classmethod